fix: extract shared HAL parser and remove dead if-else branches in API layer
