+++ src/api/admins/editAgencyAdminData.ts
@@ -30,12 +30,7 @@ export const editAgencyAdminData = async (id: string, formData: AdminData): Prom
                 twoFactorAuth,
             }),
         })
-            .then((response) => {
-                if (response.status === 200) {
-                    return response.json();
-                }
-                return response.json();
-            })
+            .then((response) => response.json())
             // eslint-disable-next-line no-underscore-dangle
             .then((data: { _embedded: CounselorData }) => data?._embedded)
     );
