+++ src/hooks/useAddOrUpdateTenantAdmin.hook.ts
@@ -4,6 +4,7 @@ import { tenantAdminsEndpoint } from '../appConfig';
 import { CounselorData } from '../types/counselor';
 import { TENANT_QUERY_KEY } from './useSingleTenantData';
 import { TENANT_ADMIN_QUERY_KEY, TENANT_ADMINS_QUERY_KEY, useTenantUserAdminData } from './useTenantUserAdminData';
+import { parseHalResponse } from '../utils/parseHalResponse';
 
 interface UseAddOrUpdateTenantAdminOptions
     extends UseMutationOptions<CounselorData, Error, CounselorData, Error | Response> {
@@ -13,29 +14,6 @@ interface UseAddOrUpdateTenantAdminOptions
 export const useAddOrUpdateTenantAdmin = ({ id, ...options }: UseAddOrUpdateTenantAdminOptions) => {
     const queryClient = useQueryClient();
     const { data } = useTenantUserAdminData({ id, enabled: !!id && id !== 'add' });
-
-    const normalizeSuccessfulResponse = async (response: unknown) => {
-        if (response instanceof Response) {
-            if (response.status === 204) {
-                return null;
-            }
-
-            const contentType = response.headers.get('content-type') || '';
-            // The backend returns HAL responses (application/hal+json), so match any JSON content
-            // type rather than the exact "application/json".
-            if (!contentType.includes('json')) {
-                return null;
-            }
-
-            const textBody = await response.clone().text();
-            if (!textBody.trim()) {
-                return null;
-            }
-
-            return JSON.parse(textBody);
-        }
-        return response;
-    };
 
     return useMutation(
         async (formData) => {
@@ -63,7 +41,7 @@ export const useAddOrUpdateTenantAdmin = ({ id, ...options }: UseAddOrUpdateTena
                 bodyData,
             });
 
-            const normalizedResponse = await normalizeSuccessfulResponse(response);
+            const normalizedResponse = await parseHalResponse(response);
             if (normalizedResponse && typeof normalizedResponse === 'object' && '_embedded' in normalizedResponse) {
                 // eslint-disable-next-line no-underscore-dangle
                 const embeddedData = (normalizedResponse as { _embedded: CounselorData })._embedded;
